requirements.txt
matplotlib~=3.7.1
numpy~=1.24.2
pytest~=7.2.2
plotly~=5.13.1
scikit-image~=0.20.0
scikit-learn~=1.2.1
scipy~=1.10.1
Pillow~=9.4.0